Increase timeout if minting fails
Since we are waiting for confirmation, it is possible for a tx to be
submitted, but then receive an error while waiting for this
confirimation. In this case, we need to wait sufficiently long so the
transaction has time to be finalized. The average duration on solana for
finalization is 15 seconds supposedly, so 1 minute should provide
sufficient time for this.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>