Change solana destination address encoding
Provide the wallet master address instead of the ATA directly. The
expected ATA is derived from the master address provided during
processing.

To remain backwars compatible, reshuffle the check to see if we've
already minted to the beginning of the mint function (and mint sign
function in cosigners), and add a function to check if this tx has
already been refunded after that. Especially the latter is important
since previously refunded txes would now be valid, causing the check to
see if the refund already happened to not happen as that lives
specifically in the refund handler, which would not be invoked in this
case.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>